Ser/De archive from envelope
